Prod release 18.0

+++ omod/src/main/java/org/openmrs/web/controller/provider/ProviderAttributeTypeFormController.java
@@ -9,28 +9,29 @@
  */
 package org.openmrs.web.controller.provider;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 import org.openmrs.ProviderAttributeType;
 import org.openmrs.api.APIException;
 import org.openmrs.api.ProviderService;
 import org.openmrs.api.context.Context;
 import org.openmrs.customdatatype.CustomDatatypeUtil;
+import org.openmrs.validator.ProviderAttributeTypeValidator;
+import org.openmrs.web.ShowFormUtil;
 import org.openmrs.web.WebConstants;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
 import org.springframework.util.StringUtils;
-import org.springframework.validation.BindException;
-import org.springframework.validation.Errors;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.SimpleFormController;
 import org.springframework.web.servlet.view.RedirectView;
 
 /**
@@ -38,95 +39,113 @@ import org.springframework.web.servlet.view.RedirectView;
  *
  * @since 1.9
  */
-public class ProviderAttributeTypeFormController extends SimpleFormController {
-	
+@Controller
+@RequestMapping(value = "admin/provider/providerAttributeType.form")
+public class ProviderAttributeTypeFormController {
+
+	private static final String FORM_VIEW = "/module/legacyui/admin/provider/providerAttributeTypeForm";
+	private static final String SUBMIT_VIEW = "providerAttributeType.list";
+
 	/** Logger for this class and subclasses */
-	protected final Log log = LogFactory.getLog(getClass());
+	private static final Logger log = LoggerFactory.getLogger(ProviderAttributeTypeFormController.class);
 	
 	/**
-	 * The onSubmit function receives the form/command object that was modified by the input form
-	 * and saves it to the db
+	 * The onSubmit function receives the form/command object that was modified by
+	 * the input form and saves it to the db
 	 *
 	 * @see org.springframework.web.servlet.mvc.SimpleFormController#onSubmit(javax.servlet.http.HttpServletRequest,
 	 *      javax.servlet.http.HttpServletResponse, java.lang.Object,
 	 *      org.springframework.validation.BindException)
 	 */
-	protected ModelAndView onSubmit(HttpServletRequest request, HttpServletResponse response, Object obj,
-	        BindException errors) throws Exception {
+	@PostMapping
+	public ModelAndView processSubmit(HttpServletRequest request,
+			@ModelAttribute("providerAttributeType") ProviderAttributeType providerAttributeType, BindingResult errors, ModelMap map)
+			throws Exception {
+
+		new ProviderAttributeTypeValidator().validate(providerAttributeType, errors);
+		
+		if (errors.hasErrors()) {
+			if (log.isDebugEnabled()) {
+				log.debug("Data binding errors: {}", errors.getErrorCount());
+			}
+			getModelMap(map);
+			return new ModelAndView(FORM_VIEW, map);
+		}
 		
 		HttpSession httpSession = request.getSession();
-		
-		String view = getFormView();
-		
+
+		String view = FORM_VIEW;
+
 		if (Context.isAuthenticated()) {
-			ProviderAttributeType providerAttributeType = (ProviderAttributeType) obj;
 			ProviderService providerService = Context.getProviderService();
-			
+
 			if (request.getParameter("save") != null) {
 				providerService.saveProviderAttributeType(providerAttributeType);
-				view = getSuccessView();
+				view = SUBMIT_VIEW;
 				httpSession.setAttribute(WebConstants.OPENMRS_MSG_ATTR, "ProviderAttributeType.saved");
 			}
 
 			// if the user is retiring out the ProviderAttributeType
 			else if (request.getParameter("retire") != null) {
 				String retireReason = request.getParameter("retireReason");
-				if (providerAttributeType.getProviderAttributeTypeId() != null && !(StringUtils.hasText(retireReason))) {
+				if (providerAttributeType.getProviderAttributeTypeId() != null
+						&& !(StringUtils.hasText(retireReason))) {
 					errors.reject("retireReason", "general.retiredReason.empty");
-					return showForm(request, response, errors);
+					return ShowFormUtil.showForm(errors, FORM_VIEW);
 				}
-				
+
 				providerService.retireProviderAttributeType(providerAttributeType, retireReason);
 				httpSession.setAttribute(WebConstants.OPENMRS_MSG_ATTR, "ProviderAttributeType.retiredSuccessfully");
-				
-				view = getSuccessView();
+
+				view = SUBMIT_VIEW;
 			}
 
 			// if the user is purging the providerAttributeType
 			else if (request.getParameter("purge") != null) {
-				
+
 				try {
 					providerService.purgeProviderAttributeType(providerAttributeType);
 					httpSession.setAttribute(WebConstants.OPENMRS_MSG_ATTR, "ProviderAttributeType.purgedSuccessfully");
-					view = getSuccessView();
-				}
-				catch (DataIntegrityViolationException e) {
+					view = SUBMIT_VIEW;
+				} catch (DataIntegrityViolationException e) {
 					httpSession.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, "error.object.inuse.cannot.purge");
 					view = "providerAttributeType.form?providerAttributeTypeId="
-					        + providerAttributeType.getProviderAttributeTypeId();
-				}
-				catch (APIException e) {
-					httpSession.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, "error.general: " + e.getLocalizedMessage());
+							+ providerAttributeType.getProviderAttributeTypeId();
+				} catch (APIException e) {
+					httpSession.setAttribute(WebConstants.OPENMRS_ERROR_ATTR,
+							"error.general: " + e.getLocalizedMessage());
 					view = "providerAttributeType.form?providerAttributeTypeId="
-					        + providerAttributeType.getProviderAttributeTypeId();
+							+ providerAttributeType.getProviderAttributeTypeId();
 				}
 			} else if (request.getParameter("unretire") != null) {
 				try {
 					providerService.unretireProviderAttributeType(providerAttributeType);
-					httpSession.setAttribute(WebConstants.OPENMRS_MSG_ATTR, "ProviderAttributeType.unretiredSuccessfully");
-					view = getSuccessView();
-				}
-				catch (APIException e) {
-					httpSession.setAttribute(WebConstants.OPENMRS_ERROR_ATTR, "error.general: " + e.getLocalizedMessage());
+					httpSession.setAttribute(WebConstants.OPENMRS_MSG_ATTR,
+							"ProviderAttributeType.unretiredSuccessfully");
+					view = SUBMIT_VIEW;
+				} catch (APIException e) {
+					httpSession.setAttribute(WebConstants.OPENMRS_ERROR_ATTR,
+							"error.general: " + e.getLocalizedMessage());
 					view = "providerAttributeType.form?providerAttributeTypeId="
-					        + providerAttributeType.getProviderAttributeTypeId();
+							+ providerAttributeType.getProviderAttributeTypeId();
 				}
 			}
 		}
-		
+
 		return new ModelAndView(new RedirectView(view));
 	}
 	
 	/**
-	 * This is called prior to displaying a form for the first time. It tells Spring the
-	 * form/command object to load into the request
+	 * This is called prior to displaying a form for the first time. It tells Spring
+	 * the form/command object to load into the request
 	 *
 	 * @see org.springframework.web.servlet.mvc.AbstractFormController#formBackingObject(javax.servlet.http.HttpServletRequest)
 	 */
-	protected Object formBackingObject(HttpServletRequest request) throws ServletException {
-		
+	@ModelAttribute("providerAttributeType")
+	protected Object formBackingObject(HttpServletRequest request) {
+
 		ProviderAttributeType providerAttributeType = null;
-		
+
 		if (Context.isAuthenticated()) {
 			ProviderService os = Context.getProviderService();
 			String providerAttributeTypeId = request.getParameter("providerAttributeTypeId");
@@ -134,24 +153,26 @@ public class ProviderAttributeTypeFormController extends SimpleFormController {
 				providerAttributeType = os.getProviderAttributeType(Integer.valueOf(providerAttributeTypeId));
 			}
 		}
-		
+
 		if (providerAttributeType == null) {
 			providerAttributeType = new ProviderAttributeType();
 		}
-		
+
 		return providerAttributeType;
 	}
-	
+
 	/**
 	 * @see org.springframework.web.servlet.mvc.SimpleFormController#referenceData(javax.servlet.http.HttpServletRequest,
 	 *      java.lang.Object, org.springframework.validation.Errors)
 	 */
-	protected Map<String, Object> referenceData(HttpServletRequest request, Object obj, Errors errors) throws Exception {
-		
-		Map<String, Object> map = new HashMap<String, Object>();
+	@GetMapping
+	public String initForm(ModelMap map) throws Exception {
+		getModelMap(map);
+		return FORM_VIEW;
+	}
+	
+	public void getModelMap(ModelMap map) {
 		map.put("datatypes", CustomDatatypeUtil.getDatatypeClassnames());
 		map.put("handlers", CustomDatatypeUtil.getHandlerClassnames());
-		
-		return map;
 	}
 }
